tehtävä 4.21 valmis

+++ osa4/controllers/blogs.js
@@ -24,11 +24,14 @@ blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
 
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log("decoded Token: ", decodedToken)
 
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
   const user = await User.findById(decodedToken.id)
+  console.log("user: ", user)
+
   try {
     if (!user) {
       const blog = new Blog({
@@ -62,8 +65,23 @@ blogsRouter.post('/', async (request, response, next) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (user._id.toString() === blog.user.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+
+  }
+  else {
+    return response.status(401).json({ error: 'Cannot delete other users blog' })
+  }
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
